Move more vote methods and tests
